Adds expanded shadows to figmaStylesParser.
