+++ src/style-dictionary/parsers/common.ts
@@ -84,3 +84,57 @@ export function getTypeFromName(name: string[]): TokenType | undefined {
     )}`
   )
 }
+
+/* These functions were an attempt to expand shadows just before entering the StyleDictionary pipeline,
+  after the parsers.
+  
+  The problem with this is that I kept over writting other values on the DesignTokens object or putting
+  the expanded data inside the composite token object, and StyleDictionary won't search for tokens inside
+  tokens.
+  
+  I leave this here, for if I (or someone) ever figures out where to put the expanded shadow tokens so that
+  the composite and the expanded tokens could coexist in the DesignTokens object and StyleDictionary parse
+  both */
+
+///**
+// * Map between the keys of a shadow token and the type of the simple token
+// * representing that value
+// */
+//export const shadowTypeMap = {
+//  blur: TokenType.shadowBlur,
+//  type: TokenType.shadowType,
+//  x: TokenType.shadowOffsetX,
+//  y: TokenType.shadowOffsetY,
+//  spread: TokenType.shadowSpread,
+//  color: TokenType.color
+//}
+
+//function _expandSingleShadowToken(token : ShadowTokenSingleValue) : DesignTokens {
+//  const res : DesignTokens = {};
+//  Object.entries(token).forEach(([key, value]) => {
+//    res[key] = {
+//      value: `${value}`,
+//      type: shadowTypeMap[key],
+//    };
+//  });
+//  return res;
+//}
+
+///**
+// * Expands an object shadow token into several simple tokens
+// * @param token 
+// * @return An object whose leaf-level entries are simple design tokens
+// */
+//export function expandShadowToken(token : ShadowDesignToken) : DesignTokens {
+//  if (Array.isArray(token.value)) {
+//    // It's a shadow consisting of several values
+//    const res : DesignTokens = {};
+//    token.value.forEach((singleShadow, index) => {
+//      res[index] = _expandSingleShadowToken(singleShadow)
+//    })
+//    return res;
+//  }
+
+//  // it's a single shadow
+//  return _expandSingleShadowToken(token.value)
+//}
