+++ src/style-dictionary/types.ts
@@ -57,7 +57,12 @@ export enum TokenType {
   motionDuration = 'motionDuration',
   motionEasing = 'motionEasing',
   ratio = 'ratio',
-  size = 'size',
+  shadowBlur = 'blur',
+  shadowOffsetX = 'x',
+  shadowOffsetY = 'y',
+  shadowSpread = 'spread',
+  shadowType = 'type',
+  size = 'size',  
   space = 'space',
   textTransform = 'textTransform',
   textDecoration = 'textDecoration',
